Rename transformers folder
