+++ rviz_default_plugins/src/rviz_default_plugins/displays/pointcloud/transformers/axis_color_pc_transformer.hpp
@@ -28,8 +28,10 @@
  */
 
 
-#ifndef RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_TRANSFORMERS__AXIS_COLOR_PC_TRANSFORMER_HPP_
-#define RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_TRANSFORMERS__AXIS_COLOR_PC_TRANSFORMER_HPP_
+#ifndef \
+  RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_TRANSFORMERS__AXIS_COLOR_PC_TRANSFORMER_HPP_
+#define \
+  RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_TRANSFORMERS__AXIS_COLOR_PC_TRANSFORMER_HPP_
 
 #include <vector>
 #include <string>
@@ -85,4 +87,4 @@ private:
 
 }  // end namespace rviz_default_plugins
 
-#endif  // RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_TRANSFORMERS__AXIS_COLOR_PC_TRANSFORMER_HPP_
+#endif  // RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_TRANSFORMERS__AXIS_COLOR_PC_TRANSFORMER_HPP_
