+++ rviz_default_plugins/src/rviz_default_plugins/displays/pointcloud/transformers/intensity_pc_transformer.cpp
