+++ rviz_default_plugins/src/rviz_default_plugins/displays/pointcloud/transformers/intensity_pc_transformer.hpp
@@ -27,8 +27,10 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_TRANSFORMERS__INTENSITY_PC_TRANSFORMER_HPP_
-#define RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_TRANSFORMERS__INTENSITY_PC_TRANSFORMER_HPP_
+#ifndef \
+  RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_TRANSFORMERS__INTENSITY_PC_TRANSFORMER_HPP_
+#define \
+  RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_TRANSFORMERS__INTENSITY_PC_TRANSFORMER_HPP_
 
 #include "rviz_common/properties/editable_enum_property.hpp"
 #include "rviz_common/properties/bool_property.hpp"
@@ -81,4 +83,4 @@ private:
 
 }  // end namespace rviz_default_plugins
 
-#endif  // RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_TRANSFORMERS__INTENSITY_PC_TRANSFORMER_HPP_
+#endif  // RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_TRANSFORMERS__INTENSITY_PC_TRANSFORMER_HPP_
