+++ rviz_default_plugins/src/rviz_default_plugins/displays/pointcloud/transformers/xyz_pc_transformer.hpp
@@ -27,8 +27,10 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_TRANSFORMERS__RGB8_PC_TRANSFORMER_HPP_
-#define RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_TRANSFORMERS__RGB8_PC_TRANSFORMER_HPP_
+#ifndef \
+  RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_TRANSFORMERS__XYZ_PC_TRANSFORMER_HPP_
+#define \
+  RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_TRANSFORMERS__XYZ_PC_TRANSFORMER_HPP_
 
 #include <vector>
 #include <string>
@@ -41,18 +43,18 @@
 namespace rviz_default_plugins
 {
 
-class RGB8PCTransformer : public rviz_default_plugins::PointCloudTransformer
+class XYZPCTransformer : public PointCloudTransformer
 {
 public:
-  unsigned char supports(const sensor_msgs::msg::PointCloud2::ConstSharedPtr & cloud) override;
+  uint8_t supports(const sensor_msgs::msg::PointCloud2::ConstSharedPtr & cloud) override;
 
   bool transform(
     const sensor_msgs::msg::PointCloud2::ConstSharedPtr & cloud,
-    unsigned int mask,
+    uint32_t mask,
     const Ogre::Matrix4 & transform,
-    rviz_default_plugins::V_PointCloudPoint & points_out) override;
+    V_PointCloudPoint & points_out) override;
 };
 
 }  // end namespace rviz_default_plugins
 
-#endif  // RVIZ_DEFAULT_PLUGINS__POINT_CLOUD_TRANSFORMERS__RGB8_PC_TRANSFORMER_HPP_
+#endif  // RVIZ_DEFAULT_PLUGINS__DISPLAYS__POINTCLOUD__POINT_CLOUD_TRANSFORMERS__XYZ_PC_TRANSFORMER_HPP_
